Yogurt7

+++ src/main/java/com/connext/wms/api/dto/GoodsDTO.java
@@ -2,6 +2,8 @@ package com.connext.wms.api.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @Author: Yogurt7_
  * @Date: 2019/1/10 14:55
@@ -10,9 +12,9 @@ import lombok.Data;
 public class GoodsDTO {
     private String sku;
     private String goodsName;
-    private Float goodsPrice;
+    private BigDecimal goodsPrice;
 
-    public GoodsDTO(String sku, String goodsName, Float goodsPrice) {
+    public GoodsDTO(String sku, String goodsName, BigDecimal goodsPrice) {
         this.sku = sku;
         this.goodsName = goodsName;
         this.goodsPrice = goodsPrice;
